refactor: cleaning & restructuring
