+++ .gitignore
@@ -1,27 +1,25 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
+#.gitignore
+pull_request_list.txt
+
 #Junk
 .vscode/
 .idea/
-
 # dependencies
 /node_modules
 /.pnp
 .pnp.js
-
 # testing
 /coverage
-
 # production
 /build
-
 # misc
 .DS_Store
 .env.local
 .env.development.local
 .env.test.local
 .env.production.local
-
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
