--- src/pages/HomePage/HomePage.jsx
@@ -1,11 +1,0 @@
-import Home from 'components/Home';
-
-const HomePage = () => {
-  return (
-    <>
-      <Home />
-    </>
-  );
-};
-
-export default HomePage;
