+++ src/redux/auth/initialState.auth.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const initialState = () => {
+    return (
+        <div>initialState</div>
+    )
+}
