+++ src/redux/auth/slice.auth.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const authSlice = () => {
+    return (
+        <div>slice</div>
+    )
+}
