+++ src/redux/auth/thunk.auth.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const thunk = () => {
+    return (
+        <div>thunk</div>
+    )
+}
