--- src/redux/dairy-calendar/dairy-calendar-selectors.js
@@ -1,1 +1,0 @@
-export const getDay = ({ dairyCalendar }) => dairyCalendar.date;
