--- src/redux/product-search/search-selectors.js
@@ -1,4 +1,0 @@
-export const getProduct = ({ product }) => product.items;
-export const getSearchLoading = ({ product }) => product.loading;
-export const getSearchError = ({ product }) => product.error;
-export const getMessage = ({ product }) => product.message;
