+++ src/redux/store.js
@@ -11,11 +11,6 @@ import {
 } from 'redux-persist';
 import storage from 'redux-persist/lib/storage';
 
-import auth from 'redux/auth/auth-slice';
-import getProductSlice from 'redux/product-search/search-slice';
-import dailyRateSlice from 'redux/daily-rate/daily-rate-slice';
-import daySlice from 'redux/day/day-slice';
-import dairyCalendarSlice from 'redux/dairy-calendar/dairy-calendar-slice';
 
 const persistConfig = {
   key: 'auth-sid',
@@ -23,15 +18,11 @@ const persistConfig = {
   whitelist: ['sid', 'accessToken', 'refreshToken'],
 };
 
-const persistedReducer = persistReducer(persistConfig, auth);
+// const persistedReducer = persistReducer(persistConfig, auth);
 
 export const store = configureStore({
   reducer: {
-    auth: persistedReducer,
-    dailyRate: dailyRateSlice,
-    product: getProductSlice,
-    dayProduct: daySlice,
-    dairyCalendar: dairyCalendarSlice,
+    // auth: persistedReducer,
   },
   middleware: getDefaultMiddleware =>
     getDefaultMiddleware({
